refactoring in place
