+++ origin-os-api/src/routes/stats.routes.js
@@ -1,6 +1,6 @@
 import express from "express";
-import { requireAuth } from "../middleware/auth.js";
-import Artwork from "../models/Artwork.js";
+import { requireAuth } from "../core/middleware/auth.js";
+import Artwork from "../modules/artworks/Artwork.js";
 
 const router = express.Router();
